specialists_service/pyproject.toml
[tool.poetry]
name = "barbershop"
version = "0.1.0"
description = "Level up task"
authors = ["Yauheni Mukhin <evgen2email@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = {extras = ["standart"], version = "^0.32.0"}
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
python-multipart = "^0.0.12"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
httpx = "^0.27.2"
pytest = "^8.3.3"
pytest-asyncio = "0.21.2"
pytest-cov = "^6.0.0"
pre-commit = "^4.0.1"
isort = "^5.13.2"
python-json-logger = "^2.0.7"
aio-pika = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
